feat(center): 기관 아이디, 유저 아이디 기준 조회 기능 추가
