+++ src/main/java/com/somemore/center/repository/NEWCenterRepository.java
@@ -2,7 +2,14 @@ package com.somemore.center.repository;
 
 import com.somemore.center.domain.NEWCenter;
 
+import java.util.Optional;
+import java.util.UUID;
+
 public interface NEWCenterRepository {
 
     NEWCenter save(NEWCenter center);
+
+    Optional<NEWCenter> findById(UUID id);
+
+    Optional<NEWCenter> findByUserId(UUID userId);
 }
